feat: removed early/absentee voting page and expanded early/abenstee voting section in elections page
