+++ packages/sdk/src/graphql/gql.ts
@@ -22,6 +22,7 @@ type Documents = {
     "\n  query GetCommunity(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    community(where: { slug: $slug }) {\n      ...PageBody\n      ...HeroImage\n      ...PageMap\n      boards {\n        ...PageList\n      }\n      topics {\n        ...TopicFields\n      }\n      documents {\n        ...DocumentList\n      }\n      actions {\n        ...ActionList\n      }\n      services {\n        ...ServiceList\n      }\n      contacts {\n        ...ContactList\n      }\n      districts {\n        ...DistrictList      \n      }\n    }\n\n    publicNotices(\n      where: { communities: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n": typeof types.GetCommunityDocument,
     "\n  query GetOrgUnit(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    orgUnit(where: { slug: $slug }) {\n      ...PageBody\n      ...HeroImage\n      actions {\n        ...ActionList\n      }\n      documents {\n        ...DocumentList\n      }\n      topics {\n        ...TopicFields\n      }\n      children {\n        ...ChildrenOrgUnits\n      }\n      contacts {\n        ...ContactList\n      }\n      parent {\n        ...OrgUnitFields\n      }\n      services {\n        ...ServiceList\n      }\n    }\n    publicNotices(\n      where: { orgUnits: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n  ": typeof types.GetOrgUnitDocument,
     "\n    query GetAssemblyDistrict(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    assemblyDistrict(where: { slug: $slug }) {\n      ...HeroImage\n      ...PageBody      \n      documents {\n        ...DocumentList\n      }\n      actions {\n        ...ActionList\n      }\n      topics {\n        ...TopicFields\n      }\n      contacts {\n        ...ContactList\n      }\n      \n      ...AssemblyMemberInfo\n      address {\n        ...AddressFields\n      }\n    }\n\n    publicNotices(\n      where: { assemblyDistricts: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n": typeof types.GetAssemblyDistrictDocument,
+    "\n  fragment GetAbsenteeVotingInfo on Query {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n": typeof types.GetAbsenteeVotingInfoFragmentDoc,
     "\n  fragment CandidateFilingInfo on Election {\n    title\n    candidates {\n      id\n      ...DocumentLink\n    }\n    electionDate\n    officesToBeFilled\n    candidateFilingStartDate\n    candidateFilingDeadline\n    candidateFilingDocuments {\n      id\n      ...DocumentLink\n    }\n  }\n": typeof types.CandidateFilingInfoFragmentDoc,
     "\n  fragment ElectionContact on ElectionsPage {\n    boroughElectionContact {\n      name\n      phone\n      email\n    }\n  }\n": typeof types.ElectionContactFragmentDoc,
     "\n  fragment ElectionPageHeader on ElectionsPage {\n    howElectionsWork\n    title\n    description\n  }": typeof types.ElectionPageHeaderFragmentDoc,
@@ -34,8 +35,7 @@ type Documents = {
     "\n  fragment ElectionResult on Election {\n    title\n    electionDate\n    result {\n      document {\n        title\n        ...DocumentLink\n      }\n      isOfficial\n    }\n  }\n": typeof types.ElectionResultFragmentDoc,
     "\n    fragment ElectionResults on Query {\n      electionResults {\n        ...ElectionResultsList\n      } \n    }\n": typeof types.ElectionResultsFragmentDoc,
     "\n  fragment UpcomingElectionDetails on Election {\n    electionDate\n    candidateFilingStartDate\n    candidateFilingDeadline\n    earlyVotingStartDate\n    absenteeApplicationDeadline\n    voterRegistrationDeadline\n\n    electionBrochure {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n\n    electionBallots {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n\n    propositions {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n\n    candidates {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n  }\n": typeof types.UpcomingElectionDetailsFragmentDoc,
-    "\n  query getAbsenteeVotingInfo {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n": typeof types.GetAbsenteeVotingInfoDocument,
-    "\n  query GetElections {\n    ...ElectionResults\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 2, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n": typeof types.GetElectionsDocument,
+    "\n  query GetElections {\n    ...ElectionResults\n    ...GetAbsenteeVotingInfo\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 1, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n": typeof types.GetElectionsDocument,
     "\n  query GetResults($search: String) {\n    electionResults(where: {\n      OR: [\n        { election: { title: { contains: $search, mode: insensitive } } },\n        { document: { title: { contains: $search, mode: insensitive } } }\n        { document: { description: { contains: $search, mode: insensitive } } }\n      ]\n    }) {\n      id\n      ...ElectionResultsList\n    }\n  }\n": typeof types.GetResultsDocument,
     "\n  query GetFacility(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    facility(where: { slug: $slug }) {\n      ...HeroImage\n      ...PageBody\n      actions {\n        ...ActionList\n      }\n      documents {\n        ...DocumentList\n      }\n\n      topics {\n        ...PageList\n      }\n      park {\n        ...PageList\n      }\n      services {\n        ...ServiceList\n      }\n      address {\n        ...AddressFields\n      }\n      contacts {\n        ...ContactList\n      }\n      hours {\n        ...HourFields\n      }\n    }\n\n    publicNotices(\n      where: { communities: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n": typeof types.GetFacilityDocument,
     "\n  query GetHomePage($take: Int, $orderBy: [PublicNoticeOrderByInput!]!) {\n    homePage {\n      id\n      title\n      description\n      heroImage\n      ...ToolbeltItems\n      ...HomePageHighlights\n    }\n\n    publicNotices(take: $take, orderBy: $orderBy) {\n      ...PublicNoticeList\n    }\n  }\n": typeof types.GetHomePageDocument,
@@ -94,6 +94,7 @@ const documents: Documents = {
     "\n  query GetCommunity(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    community(where: { slug: $slug }) {\n      ...PageBody\n      ...HeroImage\n      ...PageMap\n      boards {\n        ...PageList\n      }\n      topics {\n        ...TopicFields\n      }\n      documents {\n        ...DocumentList\n      }\n      actions {\n        ...ActionList\n      }\n      services {\n        ...ServiceList\n      }\n      contacts {\n        ...ContactList\n      }\n      districts {\n        ...DistrictList      \n      }\n    }\n\n    publicNotices(\n      where: { communities: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n": types.GetCommunityDocument,
     "\n  query GetOrgUnit(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    orgUnit(where: { slug: $slug }) {\n      ...PageBody\n      ...HeroImage\n      actions {\n        ...ActionList\n      }\n      documents {\n        ...DocumentList\n      }\n      topics {\n        ...TopicFields\n      }\n      children {\n        ...ChildrenOrgUnits\n      }\n      contacts {\n        ...ContactList\n      }\n      parent {\n        ...OrgUnitFields\n      }\n      services {\n        ...ServiceList\n      }\n    }\n    publicNotices(\n      where: { orgUnits: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n  ": types.GetOrgUnitDocument,
     "\n    query GetAssemblyDistrict(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    assemblyDistrict(where: { slug: $slug }) {\n      ...HeroImage\n      ...PageBody      \n      documents {\n        ...DocumentList\n      }\n      actions {\n        ...ActionList\n      }\n      topics {\n        ...TopicFields\n      }\n      contacts {\n        ...ContactList\n      }\n      \n      ...AssemblyMemberInfo\n      address {\n        ...AddressFields\n      }\n    }\n\n    publicNotices(\n      where: { assemblyDistricts: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n": types.GetAssemblyDistrictDocument,
+    "\n  fragment GetAbsenteeVotingInfo on Query {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n": types.GetAbsenteeVotingInfoFragmentDoc,
     "\n  fragment CandidateFilingInfo on Election {\n    title\n    candidates {\n      id\n      ...DocumentLink\n    }\n    electionDate\n    officesToBeFilled\n    candidateFilingStartDate\n    candidateFilingDeadline\n    candidateFilingDocuments {\n      id\n      ...DocumentLink\n    }\n  }\n": types.CandidateFilingInfoFragmentDoc,
     "\n  fragment ElectionContact on ElectionsPage {\n    boroughElectionContact {\n      name\n      phone\n      email\n    }\n  }\n": types.ElectionContactFragmentDoc,
     "\n  fragment ElectionPageHeader on ElectionsPage {\n    howElectionsWork\n    title\n    description\n  }": types.ElectionPageHeaderFragmentDoc,
@@ -106,8 +107,7 @@ const documents: Documents = {
     "\n  fragment ElectionResult on Election {\n    title\n    electionDate\n    result {\n      document {\n        title\n        ...DocumentLink\n      }\n      isOfficial\n    }\n  }\n": types.ElectionResultFragmentDoc,
     "\n    fragment ElectionResults on Query {\n      electionResults {\n        ...ElectionResultsList\n      } \n    }\n": types.ElectionResultsFragmentDoc,
     "\n  fragment UpcomingElectionDetails on Election {\n    electionDate\n    candidateFilingStartDate\n    candidateFilingDeadline\n    earlyVotingStartDate\n    absenteeApplicationDeadline\n    voterRegistrationDeadline\n\n    electionBrochure {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n\n    electionBallots {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n\n    propositions {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n\n    candidates {\n      id\n      title\n      file {\n        url\n      }\n      ...DocumentLink\n    }\n  }\n": types.UpcomingElectionDetailsFragmentDoc,
-    "\n  query getAbsenteeVotingInfo {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n": types.GetAbsenteeVotingInfoDocument,
-    "\n  query GetElections {\n    ...ElectionResults\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 2, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n": types.GetElectionsDocument,
+    "\n  query GetElections {\n    ...ElectionResults\n    ...GetAbsenteeVotingInfo\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 1, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n": types.GetElectionsDocument,
     "\n  query GetResults($search: String) {\n    electionResults(where: {\n      OR: [\n        { election: { title: { contains: $search, mode: insensitive } } },\n        { document: { title: { contains: $search, mode: insensitive } } }\n        { document: { description: { contains: $search, mode: insensitive } } }\n      ]\n    }) {\n      id\n      ...ElectionResultsList\n    }\n  }\n": types.GetResultsDocument,
     "\n  query GetFacility(\n    $slug: String!\n    $take: Int = 5\n    $orderDirection: OrderDirection = desc\n  ) {\n    facility(where: { slug: $slug }) {\n      ...HeroImage\n      ...PageBody\n      actions {\n        ...ActionList\n      }\n      documents {\n        ...DocumentList\n      }\n\n      topics {\n        ...PageList\n      }\n      park {\n        ...PageList\n      }\n      services {\n        ...ServiceList\n      }\n      address {\n        ...AddressFields\n      }\n      contacts {\n        ...ContactList\n      }\n      hours {\n        ...HourFields\n      }\n    }\n\n    publicNotices(\n      where: { communities: { some: { slug: { equals: $slug } } } }\n      take: $take\n      orderBy: { urgency: $orderDirection }\n    ) {\n      ...PublicNoticeList\n    }\n  }\n": types.GetFacilityDocument,
     "\n  query GetHomePage($take: Int, $orderBy: [PublicNoticeOrderByInput!]!) {\n    homePage {\n      id\n      title\n      description\n      heroImage\n      ...ToolbeltItems\n      ...HomePageHighlights\n    }\n\n    publicNotices(take: $take, orderBy: $orderBy) {\n      ...PublicNoticeList\n    }\n  }\n": types.GetHomePageDocument,
@@ -207,6 +207,10 @@ export function gql(source: "\n    query GetAssemblyDistrict(\n    $slug: String
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
+export function gql(source: "\n  fragment GetAbsenteeVotingInfo on Query {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n"): (typeof documents)["\n  fragment GetAbsenteeVotingInfo on Query {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n"];
+/**
+ * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
 export function gql(source: "\n  fragment CandidateFilingInfo on Election {\n    title\n    candidates {\n      id\n      ...DocumentLink\n    }\n    electionDate\n    officesToBeFilled\n    candidateFilingStartDate\n    candidateFilingDeadline\n    candidateFilingDocuments {\n      id\n      ...DocumentLink\n    }\n  }\n"): (typeof documents)["\n  fragment CandidateFilingInfo on Election {\n    title\n    candidates {\n      id\n      ...DocumentLink\n    }\n    electionDate\n    officesToBeFilled\n    candidateFilingStartDate\n    candidateFilingDeadline\n    candidateFilingDocuments {\n      id\n      ...DocumentLink\n    }\n  }\n"];
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
@@ -255,11 +259,7 @@ export function gql(source: "\n  fragment UpcomingElectionDetails on Election {\
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function gql(source: "\n  query getAbsenteeVotingInfo {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n"): (typeof documents)["\n  query getAbsenteeVotingInfo {\n    electionsPage {\n      stateElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      boroughElectionContact {\n        phone\n        name\n        ...ContactFields\n      }\n      earlyVotingLocations(orderBy:  {\n         order: asc\n      }) {\n        order\n        title\n        address {\n          lineOne\n          lineTwo\n          city\n          state\n          zip\n        }\n        hours {\n          id\n          day\n          open\n          close\n        }\n      }\n\n    }\n    elections(take: 1, orderBy: { electionDate: desc}) {\n      earlyVotingStartDate\n      electionDate\n      absenteeVotingApplication {\n        ...DocumentLink\n      }\n      absenteeApplicationDeadline\n    }\n  }\n"];
-/**
- * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function gql(source: "\n  query GetElections {\n    ...ElectionResults\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 2, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n"): (typeof documents)["\n  query GetElections {\n    ...ElectionResults\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 2, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n"];
+export function gql(source: "\n  query GetElections {\n    ...ElectionResults\n    ...GetAbsenteeVotingInfo\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 1, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n"): (typeof documents)["\n  query GetElections {\n    ...ElectionResults\n    ...GetAbsenteeVotingInfo\n    electionsPage {\n      heroImage\n      ...ElectionPageHeader\n      ...ElectionPollingPlaces\n      ...ElectionContact\n      ...ElectionOfficialContact\n    }\n    elections(take: 1, orderBy:  {\n      electionDate: desc\n    }) {\n      ...ElectionPageQuickLinks\n      ...UpcomingElectionDetails\n      ...ElectionVoterInformation\n      ...CandidateFilingInfo\n      ...ElectionsOfficialsInfo\n      ...ElectionResult\n    }\n  }\n"];
 /**
  * The gql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
