+++ sites/msb/components/static/Page/PageBodySection.tsx
@@ -1,4 +1,5 @@
 import clsx from 'clsx';
+import slugify from 'voca/slugify';
 
 export function PageBodySection({
   title,
@@ -15,7 +16,11 @@ export function PageBodySection({
         'not-last:mb-32 not-first:mt-32': !noMargins,
       })}
     >
-      {title && <h2 className="text-3xl font-bold mb-4">{title}</h2>}
+      {title && (
+        <h2 className="text-3xl font-bold mb-4" id={slugify(title)}>
+          {title}
+        </h2>
+      )}
       {children}
     </section>
   );
